fix(content-tooling): classify deposit slugs by DEPOSIT_RAILS, not stripped entity data
The link validator reconstructed the deposit from-/via- split by reading
src/content/input/data/exchanges — a path the content mirror strips. With
no entity data it classified every exchange as a rail, so every valid
/deposit/from-{exchange} link (coinbase, kraken, binance, …) was a
false-positive broken link — ~108 of them, jamming the whole content
pipeline's format gate.

Source the split from the same DEPOSIT_RAILS the live /deposit route uses
(extracted to a dependency-free module so the validator can import it
without content-reading side effects): in the rail set → via-, else from-.
Also skip README.md files — they're directory docs, never pages.
